+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ## v0.9.0
 
+- On `get_actions` & `get_table_deltas` only has two params (data parameters & optional parameters) both as defined as Objects.
 - On `get_actions`, `receiver` is now optional and defaults to the same value as `account`.
 - The `get_table_deltas` request was renamed to `get_table_rows`.